[common] Print styled messages, allow specifying bootfs size

+++ board/common/post_image.sh
@@ -1,6 +1,18 @@
 #!/bin/bash
 set -e
 
+TERM_BOLD=`tput smso 2>/dev/null`
+TERM_RESET=`tput rmso 2>/dev/null`
+
+function MESSAGE {
+	echo "$TERM_BOLD>>> [ACS] $1$TERM_RESET"
+}
+
+if [ -z ${BOOTFS_SIZE} ]; then
+	BOOTFS_SIZE=131072
+	BOOTFS_FAT=32
+fi
+
 COMMON_DIR="${BR2_EXTERNAL_ACS_PATH}/board/common"
 BOOTFS_DIR="${BINARIES_DIR}/bootfs"
 BOOT_DIR="${BINARIES_DIR}/bootfs/boot"
@@ -9,14 +21,9 @@ EXTLINUX_DIR="${BINARIES_DIR}/bootfs/extlinux"
 mkdir -p "${BOOT_DIR}/"
 mkdir -p "${EXTLINUX_DIR}/"
 
-if [ ! -f "${BINARIES_DIR}/zImage" ]; then
-	echo "Missing zImage - please compile the kernel first!"
-	exit 1
-fi
-
 #
 
-echo "Extracting version information"
+MESSAGE "Extracting version information"
 if [ -f "${BOOTFS_DIR}/.alpine-release" ]; then
 	export VERSION=`cat "${BOOTFS_DIR}/.alpine-release" | cut -d " " -f1 | cut -d "-" -f3`
 else
@@ -28,23 +35,26 @@ export VERSION_NAME="alpine${VERSION}-linux${LINUX}"
 
 #
 
-echo "Copying Linux kernel"
-# if compgen -G "${BINARIES_DIR}/*.dtb" > /dev/null; then
-	# cat "${BINARIES_DIR}/zImage" ${BINARIES_DIR}/*.dtb > "${BOOT_DIR}/vmlinuz-acs"
-# else
-cat "${BINARIES_DIR}/zImage" > "${BOOT_DIR}/vmlinuz-acs"
-# fi
+MESSAGE "Copying Linux kernel"
+if compgen -G "${BINARIES_DIR}/zImage.*" > /dev/null; then
+	cat ${BINARIES_DIR}/zImage.* > "${BOOT_DIR}/vmlinuz-acs"
+elif [ -f "${BINARIES_DIR}/zImage" ]; then
+	cat "${BINARIES_DIR}/zImage" > "${BOOT_DIR}/vmlinuz-acs"
+else
+	echo "Missing zImage - please compile the kernel first!"
+	exit 1
+fi
 export KERNEL="${BOOT_DIR}/vmlinuz-acs"
 
 #
 
+MESSAGE "Copying Device Tree Blobs"
 if compgen -G "${BINARIES_DIR}/*.dtb" > /dev/null; then
-	echo "Copying Device Tree Blobs"
 	mkdir -p "${BOOT_DIR}/dtbs-acs/"
 	cp ${BINARIES_DIR}/*.dtb "${BOOT_DIR}/dtbs-acs/"
 
 	if compgen -G "${BOARD_DIR}/overlays/*.dts" > /dev/null; then
-		echo "Compiling Device Tree Overlays"
+		MESSAGE "Compiling Device Tree Overlays"
 		for file in ${BOARD_DIR}/overlays/*.dts; do
 			output=${file##*/}
 			output=${output%.*}
@@ -61,14 +71,14 @@ fi
 #
 
 if [ -f "${BINARIES_DIR}/u-boot-dtb.img" ]; then
-	echo "Copying U-Boot with DTB"
+	MESSAGE "Copying U-Boot with DTB"
 	cp "${BINARIES_DIR}/u-boot-dtb.img" "${BOOTFS_DIR}/"
 fi
 
 #
 
 if [ -f "${BOARD_DIR}/uboot/boot.cmd" ]; then
-	echo "Compiling U-Boot script"
+	MESSAGE "Compiling U-Boot script"
 	if [ ! -f "${HOST_DIR}/bin/mkimage" ]; then
 		echo "mkimage not found! Please install host u-boot tools"
 		exit 1
@@ -81,13 +91,13 @@ fi
 
 #
 
-echo "Copying extlinux config"
+MESSAGE "Copying extlinux config"
 cp "${COMMON_DIR}/extlinux.conf" "${EXTLINUX_DIR}/"
 
 #
 
 if [ -f "${TARGET_DIR}/sbin/apk" ]; then
-	echo "Packing modloop"
+	MESSAGE "Packing modloop"
 	${HOST_DIR}/bin/mksquashfs \
 		"${TARGET_DIR}/lib/modules/" \
 		"${BOOT_DIR}/modloop-acs" \
@@ -98,13 +108,13 @@ if [ -f "${TARGET_DIR}/sbin/apk" ]; then
 		-Xdict-size 100%
 	export MODLOOP="${BOOT_DIR}/modloop-acs"
 
-	echo "Packing initramfs"
+	MESSAGE "Packing initramfs"
 	pushd ${TARGET_DIR}
 	find . ! -wholename "./lib/modules/*" -print0 | cpio --null -o --format=newc | gzip -c > "${BOOT_DIR}/initramfs-acs"
 	popd
 	export RAMDISK="${BOOT_DIR}/initramfs-acs"
 
-	echo "Packing apkovl"
+	MESSAGE "Packing apkovl"
 	cp "${COMMON_DIR}/alpine.apkovl.tar" "${BOOTFS_DIR}/"
 	if [ -d "${BOARD_DIR}/apkovl/" ]; then
 		tar -rf "${BOOTFS_DIR}/alpine.apkovl.tar" -C "${BOARD_DIR}/apkovl/" .
@@ -118,12 +128,18 @@ fi
 
 #
 
-echo "Packing bootfs"
+if [ -f "${BOARD_DIR}/pre_bootfs.sh" ]; then
+	. ${BOARD_DIR}/pre_bootfs.sh $*
+fi
+
+#
+
+MESSAGE "Packing bootfs"
 rm -f "${BINARIES_DIR}/bootfs.img"
 ${HOST_DIR}/sbin/mkdosfs \
-	-F 32 \
+	-F ${BOOTFS_FAT} \
 	-n bootfs \
-	-C "${BINARIES_DIR}/bootfs.img" 131072
+	-C "${BINARIES_DIR}/bootfs.img" ${BOOTFS_SIZE}
 ${HOST_DIR}/bin/mcopy \
 	-i "${BINARIES_DIR}/bootfs.img" \
 	-s \
@@ -132,6 +148,6 @@ ${HOST_DIR}/bin/mcopy \
 #
 
 if [ -f "${BOARD_DIR}/genimage.cfg" ]; then
-	echo "Running genimage"
+	MESSAGE "Running genimage"
 	${CONFIG_DIR}/support/scripts/genimage.sh -c "${BOARD_DIR}/genimage.cfg"
 fi
